+++ src/index.ts
@@ -6,6 +6,7 @@ export { default as Button } from '@/package/components/button.vue';
 export { default as Input } from '@/package/components/input.vue';
 export { default as InlineEdit } from '@/package/components/inline-edit.vue';
 export { default as SearchInput } from '@/package/components/search-input.vue';
+export { default as MentionInput } from '@/package/components/mention-input.vue';
 export { default as PasswordInput } from '@/package/components/password-input.vue';
 export { default as OtpInput } from '@/package/components/otp-input.vue';
 export { default as ColorPicker } from '@/package/components/color-picker.vue';